+++ src/HtmlToPdfTests/PageSizeTests.cs
@@ -57,7 +57,7 @@ namespace HtmlToPdfTests
                         "Parameter name: invalid"
                     };
 
-                    string expectedOutput = HelpTextGenerator.Generate(
+                    string expectedOutput = HelpTextGenerator.GenerateParserError(
                         "A required value not bound to option name is missing.",
                         string.Join(Environment.NewLine, expectedErrorMessage));
                     Assert.IsTrue(string.IsNullOrEmpty(result.StandardOutput), result.StandardOutput);